Add api-key-bundle

+++ src/Blogger/RestBundle/Controller/BlogController.php
@@ -15,15 +15,19 @@ class BlogController extends FOSRestController
      */
     public function getBlogAction(Blog $blog)
     {
-        return ['blog' => $blog];
+        if ($this->getUser()) {
+            return ['blog' => $blog];
+        }
     }
 
     public function getBlogsAction()
     {
-        $entityManager = $this->getDoctrine()->getManager();
-        $blogs = $entityManager->getRepository('BloggerBlogBundle:Blog')->findAll();
+        if ($this->getUser()) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $blogs = $entityManager->getRepository('BloggerBlogBundle:Blog')->findAll();
 
-        return ['blogs' => $blogs];
+            return ['blogs' => $blogs];
+        }
     }
 
     /**
@@ -32,21 +36,23 @@ class BlogController extends FOSRestController
      */
     public function postBlogAction(Request $request)
     {
-        $content = json_decode($request->getContent(), true);
-        $blog = new Blog();
-        $blog->setAuthor($this->getUser()->getUsername());
-        $form = $this->createForm(new BlogType(), $blog);
-        $form->submit($content);
+        if ($this->getUser()) {
+            $content = json_decode($request->getContent(), true);
+            $blog = new Blog();
+            $blog->setAuthor($this->getUser()->getUsername());
+            $form = $this->createForm(new BlogType(), $blog);
+            $form->submit($content);
 
-        if ($form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($blog);
-            $entityManager->flush();
+            if ($form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($blog);
+                $entityManager->flush();
 
-            return $this->getBlogAction($blog);
+                return $this->getBlogAction($blog);
+            }
+
+            return $form->getErrors();
         }
-
-        return $form->getErrors();
     }
 
     /**
@@ -56,20 +62,22 @@ class BlogController extends FOSRestController
      */
     public function putBlogAction(Request $request, Blog $blog)
     {
-        $content = json_decode($request->getContent(), true);
-        $entityManager = $this->getDoctrine()->getManager();
-        $blog->setAuthor($this->getUser()->getUsername());
-        $form = $this->createForm(new BlogType(), $blog);
-        $form->submit($content);
+        if ($this->getUser()) {
+            $content = json_decode($request->getContent(), true);
+            $entityManager = $this->getDoctrine()->getManager();
+            $blog->setAuthor($this->getUser()->getUsername());
+            $form = $this->createForm(new BlogType(), $blog);
+            $form->submit($content);
 
-        if ($form->isValid()) {
-            $entityManager->persist($blog);
-            $entityManager->flush();
+            if ($form->isValid()) {
+                $entityManager->persist($blog);
+                $entityManager->flush();
 
-            return $this->getBlogAction($blog);
+                return $this->getBlogAction($blog);
+            }
+
+            return $form->getErrors();
         }
-
-        return $form->getErrors();
     }
 
     /**
@@ -78,10 +86,12 @@ class BlogController extends FOSRestController
      */
     public function deleteBlogAction(Blog $blog)
     {
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($blog);
-        $entityManager->flush();
+        if ($this->getUser()) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($blog);
+            $entityManager->flush();
 
-        return $this->getBlogsAction();
+            return $this->getBlogsAction();
+        }
     }
 }
