+++ src/Blogger/RestBundle/Controller/CommentController.php
@@ -16,10 +16,12 @@ class CommentController extends FOSRestController
      */
     public function getCommentsAction(Blog $blog)
     {
-        $entityManager = $this->getDoctrine()->getManager();
-        $comments = $entityManager->getRepository('BloggerBlogBundle:Comment')->findBy(['blog' => $blog->getId()]);
+        if ($this->getUser()) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $comments = $entityManager->getRepository('BloggerBlogBundle:Comment')->findBy(['blog' => $blog->getId()]);
 
-        return ['comments' => $comments];
+            return ['comments' => $comments];
+        }
     }
 
     /**
@@ -29,22 +31,24 @@ class CommentController extends FOSRestController
      */
     public function postCommentAction(Request $request, Blog $blog)
     {
-        $content = json_decode($request->getContent(), true);
-        $comment = new Comment();
-        $comment->setBlog($blog);
-        $comment->setUser($this->getUser()->getUsername());
-        $form = $this->createForm(new CommentType(), $comment);
-        $form->submit($content);
+        if ($this->getUser()) {
+            $content = json_decode($request->getContent(), true);
+            $comment = new Comment();
+            $comment->setBlog($blog);
+            $comment->setUser($this->getUser()->getUsername());
+            $form = $this->createForm(new CommentType(), $comment);
+            $form->submit($content);
 
-        if ($form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($comment);
-            $entityManager->flush();
+            if ($form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($comment);
+                $entityManager->flush();
 
-            return $this->getCommentsAction($blog);
+                return $this->getCommentsAction($blog);
+            }
+
+            return $form->getErrors();
         }
-
-        return $form->getErrors();
     }
 
     /**
@@ -54,19 +58,21 @@ class CommentController extends FOSRestController
      */
     public function putCommentAction(Request $request, Comment $comment)
     {
-        $content = json_decode($request->getContent(), true);
-        $form = $this->createForm(new CommentType(), $comment);
-        $form->submit($content);
+        if ($this->getUser()) {
+            $content = json_decode($request->getContent(), true);
+            $form = $this->createForm(new CommentType(), $comment);
+            $form->submit($content);
 
-        if ($form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($comment);
-            $entityManager->flush();
+            if ($form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($comment);
+                $entityManager->flush();
 
-            return $this->getCommentsAction($comment->getBlog());
+                return $this->getCommentsAction($comment->getBlog());
+            }
+
+            return $form->getErrors();
         }
-
-        return $form->getErrors();
     }
 
     /**
@@ -75,10 +81,12 @@ class CommentController extends FOSRestController
      */
     public function deleteCommentAction(Comment $comment)
     {
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($comment);
-        $entityManager->flush();
+        if ($this->getUser()) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($comment);
+            $entityManager->flush();
 
-        return $this->getCommentsAction($comment->getBlog());
+            return $this->getCommentsAction($comment->getBlog());
+        }
     }
 }
